refactor(auth): 패키지 경로 수정 및 리팩토링
