+++ service/member/src/main/java/innercircle/member/application/port/in/AuthUseCase.java
@@ -1,12 +1,10 @@
 package innercircle.member.application.port.in;
 
-import innercircle.member.domain.auth.LoginRequest;
-import innercircle.member.domain.auth.LoginResponse;
+import innercircle.member.infrastructure.adapter.in.web.auth.dto.LoginRequest;
+import innercircle.member.infrastructure.adapter.in.web.auth.dto.LoginResponse;
 
 public interface AuthUseCase {
 
     LoginResponse login(LoginRequest request);
-
-
 
 }
