+++ service/member/src/main/java/innercircle/member/infrastructure/adapter/in/web/auth/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package innercircle.member.domain.auth;
+package innercircle.member.infrastructure.adapter.in.web.auth.dto;
