+++ service/member/src/test/java/innercircle/member/infrastructure/adapter/in/AuthControllerTest.java
@@ -7,8 +7,8 @@ import innercircle.global.auth.AuthErrorCode;
 import innercircle.member.application.port.in.AuthUseCase;
 import innercircle.member.application.port.out.TokenPort;
 import innercircle.member.domain.auth.LoginFailedException;
-import innercircle.member.domain.auth.LoginRequest;
-import innercircle.member.domain.auth.LoginResponse;
+import innercircle.member.infrastructure.adapter.in.web.auth.dto.LoginRequest;
+import innercircle.member.infrastructure.adapter.in.web.auth.dto.LoginResponse;
 import innercircle.member.infrastructure.adapter.out.JwtTokenAdapter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
